refactor: изменить типы аргументов у методов на Long

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -10,9 +10,9 @@ public interface FilmStorage {
 
     Film create(Film film);
 
-    Film delete(Integer filmId);
+    Film delete(Long filmId);
 
-    Film getFilmById(Integer filmId);
+    Film getFilmById(Long filmId);
 
     List<Film> getAll();
 }
